LNK-170 Added redis docker container into RedisSpecification
This commit brings in another container into the testing infrastructure. Currently it is spun up from within th RedisSpecification when the RUN_WITH_TEST_CONTAINERS is set to Y. It's disconnected from the other containers (in it's own network), as it's currently not required that the other docker containers can see this one.

+++ src/main/groovy/com/menome/util/Redis.groovy
@@ -1,34 +1,48 @@
 package com.menome.util
 
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 import redis.clients.jedis.Jedis
 import redis.clients.jedis.JedisPool
 import redis.clients.jedis.JedisPoolConfig
 
 class Redis {
 
-    static JedisPool pool = new JedisPool(new JedisPoolConfig(), ApplicationConfiguration.getString(PreferenceType.REDIS_HOST));
+    static Logger log = LoggerFactory.getLogger(Redis.class)
 
-    static  Jedis connection() {
-        pool.getResource();
+    static JedisPool pool = createConnectionPool()
+
+    static Jedis connection() {
+        Jedis jedis = null
+        if (pool) {
+            jedis = pool.getResource()
+        }
+        jedis
     }
 
     static boolean connectionOk() {
-        boolean ok = Boolean.FALSE
-        if (ApplicationConfiguration.getString(PreferenceType.USE_REDIS_CACHE) == "Y") {
-            try {
-                Jedis jedis = connection()
-                jedis.ping()
-                ok = Boolean.TRUE
-            } catch (Exception ignored) {
-                ok = Boolean.FALSE
-            }
+        boolean ok
+        try {
+            Jedis jedis = connection()
+            jedis.ping()
+            ok = Boolean.TRUE
+        } catch (Exception ignored) {
+            ok = Boolean.FALSE
         }
         ok
     }
 
-    static void clearCache(){
+    static void clearCache() {
         Jedis connection = connection()
-        connection.flushAll()
-        connection.close()
+        if (connection) {
+            connection.flushAll()
+            connection.close()
+        }
+    }
+
+    static JedisPool createConnectionPool() {
+        log.info("Connecting to Redis Host:{} on Port:{}", ApplicationConfiguration.getString(PreferenceType.REDIS_HOST), ApplicationConfiguration.getInteger(PreferenceType.REDIS_PORT))
+        def jedisPool = new JedisPool(new JedisPoolConfig(), ApplicationConfiguration.getString(PreferenceType.REDIS_HOST), ApplicationConfiguration.getInteger(PreferenceType.REDIS_PORT))
+        jedisPool
     }
 }
